refactor: mande o componente ClienteProfileHeader para a pasta shared
